NEW NEW to this. I uploaded some changes need HELP
I made some changes to the trading bot to factor in some other possible option but now im confused and not sure it worked or not could this please be checked and corrected. I would greatly appreciate it... I hope I DID this right as like i said im new to this. Thank you.